UPBGE: Remove wrong hide_render option in physics tab
The option Invisible in physics tab was originally meant to hide physics
shape during runtime.

But RNA property hide_render is now used for another purpose. It is to
hide object when we render the scene with F12.

Then this option was not working correctly.

I choose to not replace it with another RNA property because there are
other waays to hide physics shape during runtime:

in Object tab > Viewport display, we can display the shape as wire in
the viewport and it won't be drawn during runtime as we don't draw
overlays.

If we choose viewport rendering mode, we can still hide the object using
either disable all overlays, either uncheck show in viewport option in
object tab > visibility...

Already to many options for visibility, I don't think it is needed to
add another one just for the physics shape