Fix crash when switching of windows too quickly
In GHOST_TSuccess GHOST_WindowManager::endFullScreen(void) , the removed
window is not erased from m_windows list.

This is fixing a crash when we restart game in blenderplayer fullscreen
too quickly.

Note: At exit, if we start from VisualStudio, the VisualStudio window
seems resized at the resolution of the fullscreen window.

test file:
http://pasteall.org/blend/index.php?id=53047

steps to reproduce the bug before this commit:

Start the game in Blenderplayer fullscreen, and type R several times to
restart the game (restarting the game closes the current fullscreen
window and open a new fullscreen window)