UPBGE: Clarify the code
Previously, we used canvas->GetARegion() to know if we were in
blenderplayer or not. I replace with a bool.
RAS_ICanvas::IsBlenderPlayer() returns true if it is a GPG_Canvas and
false if it is a KX_BlenderCanvas.

Note: I had to remove some code in InitBlenderContextVariables, but i
was not sure of the logic of this code and i didn't notice bug without
this code then for now, i leave it like that