UPBGE: Uniformize the way we get the Main during ge runtime
I think we should almost always use the main we have set with
CTX_data_main_set(C) except maybe when we deal with wmWindowManager and
wmWindow.

I'm not sure for now but maybe we have to keep the G_MAIN unchanged
because it is used for example in wm_window_reset_drawable(void) but we
have maybe to only change the bContext main.

This commit breaks game actuator restart/load .blend in blenderplayer,
but as it's anyway not working properly, it doesn't really matter for
now.

We can try things to set correctly the bContext both in
BL_KetsjiEmbedStart and GPG_Ghost, and if needed, we can revert this
commit or part of this commit