UPBGE: Remove usage of G_MAIN and G.main in ge folder
The idea is to set the bContext G_MAIN only in BL_KetsjiEmbedStart and
GPG_ghost. After that, each time we must get the bmain, we use
CTX_data_main(C). Like that there is only 1 place to set the bmain and
only 1 method to get the bmain.

As I was changing some code in several files, I begun to "glue" the
includes in these files to prepare a future "make format".

Depending on the file, I propose to create 3 or
4 include blocks:

1. The class include (in BL_Action.cpp, the first "block" will be
BL_Action.h for example)
2. (optional) if there are includes from std between quotes, group them
(for example, include <iostream>, include <algorithm>...)
3. The blender files includes "block" with BKE, BLI, DNA, GPU....
4. The Game engine includes "block" with KX, RAS...

When we do this stuff, if there are linking errors, we can add:

ifdef __cplusplus
extern "C" {
endif....

in the concerned C file