UPBGE: Fix relative path when using startGame.
Previously the 0.0.7 the G.main->name was set to the name
of the next loaded game when using the python function startGame.
This was removed in the refactor of the launchers and caused that
image and other data loading using G.main->name to resolve
relative path were not working anymore.
To fix this issue the change of G.main is restored.

Fix issue: #563.

Initially commited by panzergame