UPBGE: Fix vehicle constraint deletion.
Previously when a controller was removed we checked vehicle constraints
using it to be removed as we do for other kind of constraints. But the
way we removed the constraint was to iterate over all vehicle constraint
and when the chassis match the controller to delete, then we call the
function RemoveVehicle which actually does the deletion of the
constraint and removed from a vehicle constraints list.
But as the elements of the list could be remove the iteration over all
the elements without taking in account the deletion of the items
resulted in invalid memory access.

To solve this issue the function RemoveVehicle is overriden to support
CcdPhysicsController as first argument and does the lookup into it with
a proper managing of iteration over variable size list.

Fix a part of #593.

Initially commited by panzergame