UPBGE: Fix unormalized vehicle axis direction.
Previously the user was able to specify a unormalized axis direction for
wheel and suspension in a vehicle wrapper, but bullet doesn't handled it
properly and generated odd wheel angles.

To fix this issue we ensure that the axis direction are well normalized.

Fix second part of issue: #593.

Initially commited by panzergame