UPBGE: Fix for loading SDL Controller gamecontrollerdb.txt file
Currently we were trying to load gamecontrollerdb.txt directly from
blender/blenderplayer folder instead of 2.8X/datafiles/gamecontroller
folder.

Now we fix this as we keep updated the gamecontrollerdb.txt file that is
in 2.8X/datafiles/gamecontroller folder.

Additionally, we remove DEV_JoystickMappingdb.h file as is outdated and
in the case that the gamecontrollerdb.txt can't be loaded then the SDL
internal database will be used.