UPBGE: Fix ping pong actions.
Previously while playing a ping pong action if the current action frame
was out of the frame interval (start and end frame) we swapped the start
and end frame to play the action in opposite wise the next time.
But this isn't enought as the current action frame was detect out of the
interval and not reset to a border of this interval. At the end it
results in playing frames after end frame or before start frame and it
also slowed the animation as some frames were lost out of the interval.

To solve this issue the current action frame is reset to the end frame
before frame swap when we detect it out of the interval.

Initially commited by panzergame