UPBGE: Fix duplication of action actuator.
Before the action actuator a member m_lastUpdate was used in the update
process of the actuator, when the actuator was duplicated this member
was set to -1 as for the constructor of the actuator.
The refactor removed this member because it was redundant, but nothing
compensated it during the duplication of the actuator.
To do so the flag of the actuator is reset to 0 excepted for continue
flag, as again the actuator constructor.

In the same time unused members of the action actuator were removed
and the python attribute blendTime which was inffective since BFBGE
was removed.

Initially commited by panzergame