UPBGE: Prevent draw request when DRW_state_image_render
Actually, i noticed that in wanderer scene there was a little flickering
which didn't happen in blender official.

This seems to fix the issue. (in bge render loop, we never set the
DRW_state to DRW_state_image_render (we could to avoid to have infinite
taa maybe, but this would need to check))