UPBGE: Implement python console.
Previously the user was not able to interact with the python API of BGE
or his python scripts to debugging his game.

This wants can be answered by a python interpreter in console started
when a shortcut is pressed.

The python interpreter use the the one from python in module "code".
When the user press the wnated shortcut the script release/scripts/bge/
interpreter.py is called. This script check if a console exist, setup
autocompletion and import bge as default. Will this script is running
the BGE does nothing. In other words the game is paused when the user
launch a python interpreter. This default behaviour was keept to avoid
obligate the user to stop manually the game if too many actions are
currently running, also a python debugging is much better when nothing
else run in the same time.

Then the user did everything he wants in the interpreter he can quit it
by pressing for linux Ctrl + D and for windows Ctrl + Z + Enter. In
windows the owner console of blender is poped when the shortcut is
pressed but it mean that the BGE will also not handle the release event
of the inputs of the shortcut. We then simulate they then the
interpreter end.

The user can change the shortcut, by default it's Ctrl + Alt + Shif + T.
Also the user can decide to allow or not python interpreter in game,
this final option is disable in all old files and as default for new
files. All these option are in UI panel Render -> UPBGE Python Console.

The python interpreter has autocompletion as describe before. It has
also persistent python locals and command history for shared beetween
each interpreter sessions. Finally it imports as default the bge module,
scene in old version was also predefined but is not a good option as if
the user use multiple scen the predefined scene will be the last scene.

Committed previously by panzergame.