UPBGE: Try to fix set object color
There's the same weird thing about object color than previously...
It needs an ID_RECALC_TRANSFORM notifier to be updated when nothing is
moving.
Then I restore this hack even if I don't understand why this is needed.

(When we play the object color action in the timeline, it plays
correcly, but in bge runtime, we need this notifier. Sybren iirc did a
commit to remove ID_RECALC_TRANSFORM for object color update few monthes
ago. It worked well, but now, for bge we need this notifier...)