Try to do modifs requested by lordloki
So as discussed on discord, I inserted code for each case and a break.

I tried to remove the else { in the code I added to do as in your
review, but I didn't notice different behaviours when we check
"continue" or not.

The code as it is now is ok for me, but if you notice something wrong,
feel free to commit directly in the branch and merge.

And thanks very much for review and point the issue with missing break;