UPBGE: Fix bge.render.drawLine
It was needed to have one of the debug options enabled to see drawLines.
Actually, we need to flush debug draw even if none of these debug
options are enabled to have the possibility to use
bge.render.drawLine...

Then I moved the call to flush outside RenderDebugProperties and removed
the flag SHOW_DEBUG_PHYSICS which I created recently but is not needed
anymore.