UPBGE: Try to fix some issues with scissors and clear
This is because there were changes yesterday, which caused issues with
the previous code i tried.

Note: There is still an issue in 1 of my test file using 2D filters, but
other files with 2D filters work. The code i commit here is what seems
to have the less issue with my test files when we use either "use
viewport render" option or not