UPBGE: Add a comment about OB_RESTRICT_RENDER in BlenderDataConversion
Even if I removed the button from physics tab ("Invisible"),
the blenderdataconversion code "hides/suspend render visibility"
of objects which have the OB_RESTRICT_RENDER restrictflag.

Then it can be used for the same purpose than before I guess except that
the setting is in outliner only instead to be both in outliner and
physics tab.