Comment not working code related to softbodies scale
It is not working neither in upbge 0.2.5.

Instead of having something which doesn't work and leads to weird
behaviours, I think it is better to have to apply scale before starting
bge. If the scale is not applied, we can directly see it and we don't
have weird behaviours (like vertices missing or invisibled object)

It is easier to debug like that.

To sum: We need to apply scale on softbody objects before ge start for
now unless we find a fix