UPBGE: Rigidbody/Dynamics continuous collision detection (#1316)
A new option is available per object to set-up the continuous collision detection. There are 2 new properties:

MotionThreshold: Sets the delta of movement that has to happen in one physics tick to trigger the continuous motion detection.
SweptSphereRadius: The radius of the sphere that is used to check for possible collisions when ccd is actived.

Basically, the SweptSphereRadius has to be a little bit smaller than object volume to avoid false detections (i.e default cube 2x2x2, the swept radius should be 0.9).

The motionthreshold calculus is done in the following way:
For example, a bullet velocity is around 1000 m/s. If you shoot a bullet in your game (60 fps) then in each frame your bullet goes through 1000/60 = 16.66m. Therefore, you should be set your MotionThreshold around 15.

These new options will be have a penalty in the performance. Therefore, you should be use with care and in a few objects only.

Thanks a lot to @youle31 for his exhaustive review and testing.