UPBGE: Expose ERP, ERP2 and CFM bullet parameters
With this parameters we can tuning the simulations.
The ERP specifies what proportion of the joint error will be fixed
during the next simulation step.
If ERP=0 then no correcting force is applied and the bodies will
eventually drift apart as the simulation proceeds.
If ERP=1 then the simulation will attempt to fix all joint error during
the next time step.
However, setting ERP=1 is not recommended, as the joint error will not
be completely fixed due to various internal approximations.
A value of ERP=0.1 to 0.8 is recommended (0.8 is the default).

ERP2 is for contact constraints and ERP for non contact constraints.

CFM is related to soft constraints.
If CFM is set to zero, the constraint will be hard.
If CFM is set to a positive value, it will be possible to violate the
constraint by "pushing on it" (for example, for contact constraints by
forcing the two contacting objects together).
In other words the constraint will be soft, and the softness will
increase as CFM increases.
What is actually happening here is that the constraint is allowed to be
violated by an amount proportional to CFM times the restoring force that
is needed to enforce the constraint.

ERP, ERP2 and CFM values are exposed through bullet scene settings and
python functions (setERPNonContact, setERPContact and setCFM) because
they are very interesting for fine tunning the simulations.