UPBGE: Implement character maximum slope python functions
The maximum slope feature was recovered previously except the python
functions

Maximum slope is bullet feature which can be used to restrict the
player movements, this value is between 0 and pi/2, exposed
in UI and in python under KX_CharacterWrapper.maxSlope.

In the same time the PHY_ShapeProps struct is removed as its
purpose was just to store blender data used in ConvertObject
function but this function was already doing data conversion from
blender directly, instead mixing different levels, this struct is
removed.

Previously committed by panzergame