Try to fix an issue with converted objects during runtime
Previously, as converted objects during runtime were not marked as
replica objects, endObject() was not freeing the blenderObject.

This branch is an attempt to free the blenderObject for converted
objects during runtime.

The problem is that when we invoke BKE_collection_object_remove(ob, bool
free_user), if free_user is true it causes an issue when we call
BKE_id_free... idk if it matters or not.