UPBGE: Fix assert error in debug builds when removing replica
When we add a replica object, BKE_id_copy is used and it increments
newobject->id.us.

When we delete this replica, we use BKE_id_delete and an exception is
raised in debug mode to tell us that user count is == 1

To avoid to deal with users count, I inserted
id_us_min(&newob->id) just after adding the replica when we are using
normal method to add a replica.