UPBGE: Fix special case where converter was unregistering objects several times
When we unregister a pair <Object *, KX_GameObject *> from
m_map_blender_to_gameobject in scene converter, we need to ensure this
pair has not been unregistered previously.

It can happen when BlenderDataConversion decides to not convert an
object:

This is the case when a parent is in an Active layer whereas the
children are NOT in an Active layer.