UPBGE: Fix shape replacement for character.
Previously the shape replacement for character wasn't managed and the
previous shape was deleted without relink to the new shape.

To solve this issue CcdCharacter::ReplaceShape is implemented to replace
the shape, this function is called into ReplaceControllerShape.

As character could not use a non convex shape, some shape replacement
may fail (e.g with triangle mesh) the user notice it with the return
value of obj.replacePhysicsShape which is now a boolean.

Fix issue: #968.

Previously comitted by Panzergame