UPBGE: Implement angular servo control motion logic brick.
Previously the servo control logic brick was able to control only the
linear velocity using force. But some users showed that in some non-rare
case it could be needed to servo control the angular velocity via
torques.

To fill this gap, the servo control logic brick exposes a enum named
"Servo Type" to select either linear or angular velocity, linear is the
default. This option is used in KX_ObjectActuator to compute from linear
or angular velocity a force or torque.

Note: The reference object in case of angular velocity is just used to
substract its angular velocity contrary to linear velocity which compute
the linear velocity of the reference object on the position of the
affected object.

Previously committed by panzergame