UPBGE: Fix angular velocity motion servo with little objects.
Previously the force computed in servo control logic brick was
multiplied by the mass. This behaviour is correct because bullet in
btRigidBody::integrateVelocities compute the linear velocity using force
/ mass.
But the angular velocity doesn't use this formula. For the angular
velocity we have to multiply the torque by the inertia of the body
because bullet compute angular velocity with: torque * inertia_tensor.

Previously committed by panzergame