UPBGE: Implement std::vector remove and add utilities.
Two functions are added to replace frequently used behaviour:
remove if an element is found in a list, or add if an element is
not present in a list. The functions CM_ListRemoveIfFound
and CM_ListAddIfNotFound assume these goal, they are defined in
new file CM_List.h.

Previously committed by panzergame