UPBGE: Fix continuos activation for mouse actuator
This fixes issue #1352

Until now it was no necessary to active the true pulse in the sensor to
make the mouse actuator works. This was because the mouse actuator
requested continuos updates indepently of sensor pulses. Due to this
behaviour several bugs when we used several mouse actuators occured.

Now with this change the mouse actuator doesn't request continuos
updates after activation. Therefore, to make it work it will be
necessary to active the true pulse in the sensor.