UPBGE: Try to get rid of resetTaaSamples API
I noticed a slight difference between viewport drawing and custom bge
loop.

Investigating on it, i realized that the difference comes from
vedata->stl->g_data->view_update was always false contrarly to viewport.

It is because in viewport, the depsgraph detects all changes done in the
scene each time there is a notifier and notifies the drw_engines.
This behaviour comes from a callback set in wm_init_exit:

DEG_editors_set_update_cb(ED_render_id_flush_update,
ED_render_scene_update);

At a lower level, the drw_engine is notified by:

DRW_notify_view_update(const DRWUpdateContext *update_ctx)

When we are using viewport render, ARegion viewports are used, then this
function works correctly. But as we are using our own GPUViewports in
custom bge loop, this function is not working correctly.

This commit is a first attempt to have this function working even if we
are using our own custom GPUViewports.

If we have it working, we can normally get rid of resetTaaSamples API as
the depsgraph automatically triggers the needed functions to set the
taa_current_sample to 1.

TODOS:
- Check ImageRender (no notifier is called here)
- Check Overlay scenes
- check multiple viewports