UPBGE: Prevent some code to be executed if not needed in BL_Action
If bAction has already been updated to localframe, there is no need to
run other pieces of code, then I added a bool to check if the action has
already been updated.

example:

An object has several modifiers and several nodetrees.
If the current updated action is an action on a modifier, there
is no need to execute the code related to nodetrees