UPBGE: Allow full rendered passes in ImageRender (high perf cost)
It allows for example to composite the main render pass with a full soft
shadows pass.

Or to display each frame a fully antialiased render.

If we set samples to 16 for example, the scene (COMBINED pass or other
passes like SHADOWS, MIST, AO...) will be rendered 16 times
in 1 frame (just for ImageRender) so the performances cost is very high.

See render_passes (v3d.shading.render_pass) documentation in API