UPBGE: improve text properties conversion
Instead of pass an empty string for String property creation we pass the
stream data.

Additionally, we remove the stream fail check as the conversion is well
done but we mantain the bad check to check the stream integrity. This
way we avoid the console error message that it is not very useful for
the user (as the conversion is really done).