Remove KX_PythonLogger & make KX_Scene a proxy
Revert 1d14fb085 due to request and make KX_Scene a
KX_PythonProxy instead. Note that KX_PythonProxy does
not have a default constructor yet and it still uses
py_base_new to create a proxy.

It needs to be changed in future if want to allow users to
provide a custom scene implementation like they can with
KX_GameObject.