UPBGE: Implement python fast keyword parsing.
The usage of undocumented python function
_PyArg_ParseTupleAndKeywordsFast (already used in blender) helps to
improve performance spend during function argument parsing. After some
test this new function is 2/3 of the runnign time of the previous
parsing function.

But this function has a drawback: it request more code and variable to
be used especially a python parser pre created. To avoid to write this
code each time the template function EXP_ParseTupleArgsAndKeywords is
used. This function receive the args and kwds tuples, the format, the
keywords name and pointers converted values.

Previously commited by panzergame