UPBGE: Avoid to set visibled state to init state value at ge exit
I'm not totally sure it should be like that but it avoids
visibled controller state to be set to initial state value at ge exit
(because of undo at exit).

visibled state (at least in the new code) seems to be used only in logic
design time, but when bge start, only init_state seems taken into account.
It changed since initial commit related to states:
https://developer.blender.org/rB5372def2b069699809cf9439381b03364137adb5

This commit should restore the behaviour of 0.2.5

It doesn't prevent to "reset" visibled state to initial state on .blend
loading...