UPBGE: Fix for mouse actuator continuos activation
Now it works perfectly when activating from an always sensor and mouse
movement sensor.

Now it behaves the same way than 0.2.5. Therefore it is necessary to
activate true pulse in Always sensor or Mouse movement sensor to make
the Mouse actuator work. This is due that the mouse actuator are not
pushing itself the activation.

Additionally, it introduces a boolean check for initial skipping instead
the modifying the oldposition.