UPBGE: Fix contact point list for ghost objects.
Two issues were noticed into CallbackTriggers function.
First manifolds for objects ignoring contact response were
automatically removed, apparently this was needed for older version
of bullet which didn't updated the manifold. This behaviour caused
to always send a invalid contact point list to the user in case of
ghost object collision. To solve this issue the clear is removed as
after some test the bug is not anymore present.
Secondly debug draw of contact points was proceed in CallbackTriggers
but btIDebugDraw is already managing it the same…
The manual debug draw is removed.

Fix issue #636.

Previously committed by panzergame