UPBGE: Fix free of collision data.
Previously when calling python function obj.collide(...)
the collision data PHY_ICollData was owned by noone and
so leaked in memory.

To fix this issue the owner ship if transfered to
KX_CollisionContactPointList.

Also a function KX_GameObject::HasCollisionCallbacks is
implemented to not create contact point list for nothing.

Previously committed by panzergame