UPBGE: Fix reference counting of module items.
The item passed to PyDict_SetItemString will be incref, so if the item
is constructed in the current scope and not stored in an external data,
the item must be decref.

This change from commit e9f90370ae61d9587a651461f95da8c7af88c1b1 is
recommitted as it is safe. The rest of that commit makes crash with
default scene after start-exit quickly loop.

Thanks to @youle31 for the investigation.