UPBGE: Add our own bitflag for uiBut
Due to the uiBut.flag reached to its maximum and it is completly used by
Blender defines, it is better for us to use a new bitflag.

Maybe if in the future Blender use another method to increase the
bitflag capacity we can discard our method and come back with its
method.