UPBGE: Try to fix debug build at first run with MSVC
When building upbge in debug mode from scratch with MSVC, extern_recastnavigation lib
was spawning a weird message (MC not found or can't open MC or something
like that)

It seems it comes from remove_cc_flag_unsigned_char() macro at the
beginning of recastnavigation cmakelists.

This macro is only used to build extern_recastnavigation and is defined
in macros.cmake.

It contained this for linux:

remove_cc_flag("-funsigned-char")

and this for MSVC:

remove_cc_flag("/J")

but according to this article written in 2021, there is
nothing corresponding to -funsigned-char for Visual Studio:

https://www.intel.com/content/www/us/en/docs/cpp-compiler/developer-guide-reference/2021-9/funsigned-char.html

Then I comment it