UPBGE: Fix no response objects collision not refreshed
on discord, @  musikai reported a changement of collision behaviour in a
file (i'll join it it commit comments) between a Sensor and a Static
object since 0.3.

It's because of this commit:

https://github.com/UPBGE/upbge/commit/085232d1ae6e8d73f045df8a15d9a94571b0af36

In current bullet version, part of the code removed in mentionned
commit is needed to properly handle no contact response (ghost + sensor)
objects collisions, else, collisions are not always refreshed, as shown
in first test file (i'll join it in commit comments)

The second test file prints contact points list to see if
collisionCallback when collision between a sensor object and a static
ghost object works ok. It sounds ok with current
bullet version. (issue mentionned here:
https://github.com/UPBGE/upbge/issues/636)

Then I restore the code related to no contact response objects in
CallbackTriggers