Fix: GP: Wrong offset vector when drawing onto surfaces
The offset from the surface was calculated based on only the normal
which didn't used to be the case.
The fix computes the offset based on the surface point and the direction
of the view.
Also makes sure to not draw overlays when computing the depths using
`V3D_DEPTH_NO_GPENCIL`.

(cherry picked from commit 9d5c5d4a93e218bd91d4255a21fb27c4fbcc03d8)