Fix: GP: UI errors when no layer tree node is active
The UI code assumed that either a layer or a group is always
active which is not always the case. It can happen that nothing
is active.
The fix makes sure the this case is handled correctly.

(cherry picked from commit a25456eaf1ee63291cdd75e946b399ebd4a37d87)